+++ test/gen-enums.config.ts
@@ -5,6 +5,5 @@ export default {
     server: "test/.test-output/enums/server",
     client: "test/.test-output/enums/client"
   },
-  softDeleteColumn: null,
   includeMethodsDepth: 2
 };